+++ build_runner/bin/graph_inspector.dart
@@ -50,7 +50,7 @@ Future<void> main(List<String> args) async {
   stdout.writeln('Loading asset graph at ${assetGraphFile.path}...');
 
   assetGraph = AssetGraph.deserialize(assetGraphFile.readAsBytesSync());
-  packageGraph = PackageGraph.forThisPackage();
+  packageGraph = await PackageGraph.forThisPackage();
 
   var commandRunner = CommandRunner<bool>(
       '', 'A tool for inspecting the AssetGraph for your build')