+++ build_runner_core/lib/src/asset_graph/serialization.dart
@@ -8,7 +8,7 @@ part of 'graph.dart';
 ///
 /// This should be incremented any time the serialize/deserialize formats
 /// change.
-const _version = 22;
+const _version = 23;
 
 /// Deserializes an [AssetGraph] from a [Map].
 class _AssetGraphDeserializer {
@@ -35,9 +35,19 @@ class _AssetGraphDeserializer {
 
   /// Perform the deserialization, should only be called once.
   AssetGraph deserialize() {
+    var packageLanguageVersions = {
+      for (var entry in (_serializedGraph['packageLanguageVersions']
+              as Map<String, dynamic>)
+          .entries)
+        entry.key: entry.value != null
+            ? LanguageVersion.parse(entry.value as String)
+            : null
+    };
     var graph = AssetGraph._(
-        _deserializeDigest(_serializedGraph['buildActionsDigest'] as String),
-        _serializedGraph['dart_version'] as String);
+      _deserializeDigest(_serializedGraph['buildActionsDigest'] as String),
+      _serializedGraph['dart_version'] as String,
+      packageLanguageVersions,
+    );
 
     var packageNames = _serializedGraph['packages'] as List;
 
@@ -218,6 +228,8 @@ class _AssetGraphSerializer {
       'buildActionsDigest': _serializeDigest(_graph.buildPhasesDigest),
       'packages': packages,
       'assetPaths': assetPaths,
+      'packageLanguageVersions': _graph.packageLanguageVersions
+          .map((pkg, version) => MapEntry(pkg, version?.toString())),
     };
     return utf8.encode(json.encode(result));
   }