+++ build_runner_core/lib/src/package_graph/package_graph.dart
@@ -4,6 +4,7 @@
 
 import 'dart:io';
 
+import 'package:package_config/package_config.dart';
 import 'package:path/path.dart' as p;
 import 'package:yaml/yaml.dart';
 
@@ -11,8 +12,13 @@ import '../util/constants.dart';
 
 /// The SDK package, we filter this to the core libs and dev compiler
 /// resources.
-final PackageNode _sdkPackageNode =
-    PackageNode(r'$sdk', sdkPath, DependencyType.hosted);
+final _sdkPackageNode = PackageNode(
+    r'$sdk',
+    sdkPath,
+    DependencyType.hosted,
+    // A fake language version for the SDK, we don't allow you to read its
+    // sources anyways, and invalidate the whole build if this changes.
+    LanguageVersion(0, 0));
 
 /// A graph of the package dependencies for an application.
 class PackageGraph {
@@ -53,7 +59,7 @@ class PackageGraph {
 
   /// Creates a [PackageGraph] for the package whose top level directory lives
   /// at [packagePath] (no trailing slash).
-  factory PackageGraph.forPath(String packagePath) {
+  static Future<PackageGraph> forPath(String packagePath) async {
     /// Read in the pubspec file and parse it as yaml.
     final pubspec = File(p.join(packagePath, 'pubspec.yaml'));
     if (!pubspec.existsSync()) {
@@ -64,27 +70,27 @@ class PackageGraph {
     final rootPubspec = _pubspecForPath(packagePath);
     final rootPackageName = rootPubspec['name'] as String;
 
-    final packageLocations = _parsePackageLocations(packagePath)
-      ..remove(rootPackageName);
+    final packageConfig =
+        await findPackageConfig(Directory(packagePath), recurse: false);
 
     final dependencyTypes = _parseDependencyTypes(packagePath);
 
     final nodes = <String, PackageNode>{};
-    final rootNode = PackageNode(
-        rootPackageName, packagePath, DependencyType.path,
-        isRoot: true);
-    nodes[rootPackageName] = rootNode;
-    for (final packageName in packageLocations.keys) {
-      if (packageName == rootPackageName) continue;
-      nodes[packageName] = PackageNode(packageName,
-          packageLocations[packageName], dependencyTypes[packageName],
-          isRoot: false);
+    for (final package in packageConfig.packages) {
+      var isRoot = package.name == rootPackageName;
+      nodes[package.name] = PackageNode(
+          package.name,
+          package.root.toFilePath(),
+          isRoot ? DependencyType.path : dependencyTypes[package.name],
+          package.languageVersion,
+          isRoot: isRoot);
     }
-
+    final rootNode = nodes[rootPackageName];
     rootNode.dependencies
         .addAll(_depsFromYaml(rootPubspec, isRoot: true).map((n) => nodes[n]));
 
-    final packageDependencies = _parsePackageDependencies(packageLocations);
+    final packageDependencies = _parsePackageDependencies(
+        packageConfig.packages.where((p) => p.name != rootPackageName));
     for (final packageName in packageDependencies.keys) {
       nodes[packageName]
           .dependencies
@@ -95,7 +101,8 @@ class PackageGraph {
 
   /// Creates a [PackageGraph] for the package in which you are currently
   /// running.
-  factory PackageGraph.forThisPackage() => PackageGraph.forPath('./');
+  static Future<PackageGraph> forThisPackage() =>
+      PackageGraph.forPath(p.current);
 
   /// Shorthand to get a package by name.
   PackageNode operator [](String packageName) => allPackages[packageName];
@@ -131,7 +138,10 @@ class PackageNode {
   /// Whether this node is the [PackageGraph.root].
   final bool isRoot;
 
-  PackageNode(this.name, String path, this.dependencyType, {bool isRoot})
+  final LanguageVersion languageVersion;
+
+  PackageNode(this.name, String path, this.dependencyType, this.languageVersion,
+      {bool isRoot})
       : path = _toAbsolute(path),
         isRoot = isRoot ?? false;
 
@@ -146,34 +156,6 @@ class PackageNode {
   /// `null`.
   static String _toAbsolute(String path) =>
       (path == null) ? null : p.canonicalize(path);
-}
-
-/// Parse the `.packages` file and return a Map from package name to the file
-/// location for that package.
-Map<String, String> _parsePackageLocations(String rootPackagePath) {
-  var packagesFile = File(p.join(rootPackagePath, '.packages'));
-  if (!packagesFile.existsSync()) {
-    throw StateError('Unable to generate package graph, no `.packages` found. '
-        'This program must be ran from the root directory of your package.');
-  }
-  var packageLocations = <String, String>{};
-  for (final line in packagesFile.readAsLinesSync().skip(1)) {
-    var firstColon = line.indexOf(':');
-    var name = line.substring(0, firstColon);
-    assert(line.endsWith('lib/'));
-    // Start after package_name:, and strip out trailing 'lib/'.
-    var uriString = line.substring(firstColon + 1, line.length - 4);
-    // Strip the trailing slash, if present.
-    if (uriString.endsWith('/')) {
-      uriString = uriString.substring(0, uriString.length - 1);
-    }
-    var uri = Uri.tryParse(uriString) ?? Uri.file(uriString);
-    if (!uri.isAbsolute) {
-      uri = p.toUri(p.join(rootPackagePath, uri.path));
-    }
-    packageLocations[name] = uri.toFilePath(windows: Platform.isWindows);
-  }
-  return packageLocations;
 }
 
 /// The type of dependency being used. This dictates how the package should be
@@ -212,14 +194,13 @@ DependencyType _dependencyTypeFromSource(String source) {
   throw ArgumentError('Unable to determine dependency type:\n$source');
 }
 
-/// Read the pubspec for each package in [packageLocations] and finds it's
+/// Read the pubspec for each package in [packages] and finds it's
 /// dependencies.
-Map<String, Set<String>> _parsePackageDependencies(
-    Map<String, String> packageLocations) {
+Map<String, Set<String>> _parsePackageDependencies(Iterable<Package> packages) {
   final dependencies = <String, Set<String>>{};
-  for (final packageName in packageLocations.keys) {
-    final pubspec = _pubspecForPath(packageLocations[packageName]);
-    dependencies[packageName] = _depsFromYaml(pubspec);
+  for (final package in packages) {
+    final pubspec = _pubspecForPath(package.root.toFilePath());
+    dependencies[package.name] = _depsFromYaml(pubspec);
   }
   return dependencies;
 }