+++ build_runner_core/test/asset/file_based_test.dart
@@ -12,11 +12,11 @@ import 'package:build_runner_core/build_runner_core.dart';
 
 import 'package:_test_common/common.dart';
 
-final PackageGraph packageGraph =
-    PackageGraph.forPath('test/fixtures/basic_pkg');
-final String newLine = Platform.isWindows ? '\r\n' : '\n';
+final newLine = Platform.isWindows ? '\r\n' : '\n';
 
-void main() {
+void main() async {
+  final packageGraph = await PackageGraph.forPath('test/fixtures/basic_pkg');
+
   group('FileBasedAssetReader', () {
     final reader = FileBasedAssetReader(packageGraph);